Initial general javascript client
This client uses babel and webpack to allow support for node and the browser at the same time.
Using webpack and babel allows us to simply use ES6 without worrying about browser compatibility.
The main downside of this is larger size.

It is published to npm to register the name.